+++ tests/plugins/test_database.py
@@ -69,6 +69,14 @@ async def add_note(req: Request):
     return JSONResponse(data)
 
 
+@app.route("/notes/bulk_create", methods=["POST"])
+async def bulk_create_notes(request):
+    data = await request.json()
+    query = notes.insert()
+    await request.database.executemany(query, data)
+    return JSONResponse({"notes": data})
+
+
 @app.route("/notes/{note_id:int}", methods=["GET"])
 async def note_row(req: Request):
     note_id = req.path_params["note_id"]
@@ -146,3 +154,19 @@ def test_database_isolated_during_test_cases():
         response = client.get("/notes")
         assert response.status_code == 200
         assert response.json() == [{"text": "just one note", "complete": True}]
+
+
+def test_database_executemany():
+    with TestClient(app) as client:
+        data = [
+            {"text": "buy the milk", "complete": True},
+            {"text": "walk the dog", "complete": False},
+        ]
+        response = client.post("/notes/bulk_create", json=data)
+        assert response.status_code == 200
+        response = client.get("/notes")
+        assert response.status_code == 200
+        assert response.json() == [
+            {"text": "buy the milk", "complete": True},
+            {"text": "walk the dog", "complete": False},
+        ]