+++ yast/plugins/template/__init__.py
@@ -1,0 +1,5 @@
+from yast.applications import Yast
+
+
+def plugin_init(app: Yast, config={}) -> None:
+    print("template init")